Add a helper for OIDC authorization.
This helper service eases OIDC configuration for tctl users.

It also allows tctl to retrieve an access token from providers that do not
support the device authorization flow, or cases where the device authorization
flow is disabled for an OAuth Client ID. This is done without tctl needing to
have the Client Secret, which is normally required for the Authorization flow.