Make server listen to all interfaces instead of just loopback
We should not assume request always come from the loopback interface
even for local dev environment. For example, if the server runs inside
of a container, a request originated from either the host or another
container is routed through the VM's interface instead of loopback.

This was originally reported on the forum:
https://community.temporal.io/t/running-temporal-proxy-server-in-docker/8107