Show how to use Temporal as a queue
This example shows how you can use Temporal's event history and workflow
input as a queue. It implements a batching mechanism where a workflow
continuously listens for incoming signals, as well as a timer, and
processes the batch when the timer fires, followed by restarting the
timer. The workflow keeps track of the number of incoming events and
decides to continue-as-new when there are too many events recorded in
the event history, passing the accumulated events in the next workflow's
input.