Payload converter based on payload size
This example shows how to implement a payload converter that detects
when a payload is larger than a certain threshold, in which case it
writes it to a storage system of choice instead of the temporal server.
In this particular example payloads are written to a local file but it
can be adapted to work with any storage. Such a converter is useful when
a workflow/activity regularly has payloads that risk exceeding
Temporal's hard limit of 2MB.